CREATED SECURITY LAYER

+++ notes/src/main/java/es/notes/notes/exception/NoteNotFoundException.java
@@ -3,7 +3,7 @@ package es.notes.notes.exception;
 public class NoteNotFoundException extends RuntimeException {
 
     // Constructor que recibe el ID y construye un mensaje descriptivo
-    public NoteNotFoundException(Integer id) {
+    public NoteNotFoundException(Long id) {
         super("No se pudo encontrar la nota con ID: " + id);
     }
 }
